added the find appliances by brand and display appliances by type swi…

+++ GroupAssignment1/ApplianceIO.cs
@@ -30,7 +30,7 @@ namespace GroupAssignment1
                         appliances.Add(appliance);
                         break;
                     case "2":
-                        appliance = new Vacuum(fields[0], fields[1], int.Parse(fields[2]), double.Parse(fields[3]), fields[4], double.Parse(fields[5]), fields[6], double.Parse(fields[7]));
+                        appliance = new Vacuum(fields[0], fields[1], int.Parse(fields[2]), double.Parse(fields[3]), fields[4], double.Parse(fields[5]), fields[6], int.Parse(fields[7]));
                         appliances.Add(appliance);
                         break;
                     case "3":
